+++ packages/core/src/index.ts
@@ -1,5 +1,16 @@
 export * from "./types";
 export * from "./components";
-export { addTelemetryInitializer, Envelope } from "@equinor/mad-insights";
+export { authenticateSilently } from "@equinor/mad-auth";
+export {
+    appInsightsHasBeenInitialized,
+    addTelemetryInitializer,
+    Envelope,
+    metricKeys,
+    metricStatus,
+    track,
+    trackCustom,
+    trackLongTerm,
+    trackShortTerm,
+} from "@equinor/mad-insights";
 export * from "./hooks";
-export { setConfig, setEnvironment, useMadConfig } from "./store/mad-config/mad-config";
+export * from "./store";