+++ packages/core/src/store/index.ts
@@ -1,0 +1,2 @@
+export * from "./demo-mode";
+export * from "./language";