+++ packages/core/src/utils/createMadCoreNavigator.tsx
@@ -11,6 +11,7 @@ import { ReleaseNotesScreen } from "../components/screens/release-notes/ReleaseN
 import { AboutScreen } from "../components/screens/AboutScreen";
 import { CreateIncidentScreen } from "../components/screens/CreateIncidentScreen";
 import { SelectLanguageScreen } from "../components/screens/language/SelectLanguageScreen";
+import { ServiceMessageProvider } from "../components/service-message/ServiceMessageProvider";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any -- We need to specify how a general function looks like
 type GeneralFunction = (...args: any) => any;
@@ -27,29 +28,31 @@ export const createMadCoreNavigator = <T extends ParamListBase>(
         return (
         <AppInsightsInitializer config={config.applicationInsights}>
             <EnvironmentProvider>
-                <AnnouncementsProvider>
-                    <Stack.Navigator {...props} initialRouteName={"Login"}>
-                        <Stack.Screen
-                            name="Login"
-                            component={LoginScreen}
-                            options={{headerShown: false, customSubHeaderShown: true}}
-                        />
-                        <Stack.Screen name="ReleaseNotes" component={ReleaseNotesScreen} />
-                        <Stack.Screen
-                            name="WhatsNew"
-                            component={WhatsNewScreen}
-                        />
-                            {config.about && <Stack.Screen name="About" component={AboutScreen} />}
-                            {config.serviceNow && (
-                                <Stack.Screen name="Feedback" component={CreateIncidentScreen} />
-                            )}
-                            {config.language.supportedLanguages.length > 1 && <>
-                                <Stack.Screen name="SelectLanguage" component={SelectLanguageScreen} />
-                                <Stack.Screen name="SelectLanguageOnboarding" component={SelectLanguageScreen} options={{headerBackVisible: false}} />
-                            </>}
-                            {props.children}
-                        </Stack.Navigator>
-                </AnnouncementsProvider>
+                <ServiceMessageProvider>
+                    <AnnouncementsProvider>
+                        <Stack.Navigator {...props} initialRouteName={"Login"}>
+                            <Stack.Screen
+                                name="Login"
+                                component={LoginScreen}
+                                options={{headerShown: false, customSubHeaderShown: true}}
+                                />
+                            <Stack.Screen name="ReleaseNotes" component={ReleaseNotesScreen} />
+                            <Stack.Screen
+                                name="WhatsNew"
+                                component={WhatsNewScreen}
+                                />
+                                {config.about && <Stack.Screen name="About" component={AboutScreen} />}
+                                {config.serviceNow && (
+                                    <Stack.Screen name="Feedback" component={CreateIncidentScreen} />
+                                    )}
+                                {config.language.supportedLanguages.length > 1 && <>
+                                    <Stack.Screen name="SelectLanguage" component={SelectLanguageScreen} />
+                                    <Stack.Screen name="SelectLanguageOnboarding" component={SelectLanguageScreen} options={{headerBackVisible: false}} />
+                                </>}
+                                {props.children}
+                            </Stack.Navigator>
+                    </AnnouncementsProvider>
+                </ServiceMessageProvider>
             </EnvironmentProvider>
         </AppInsightsInitializer>
         );