+++ packages/core/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @equinor/mad-core
 
+## 0.3.1
+
+### Patch Changes
+
+-   7c9c6da: export most of `mad-insights`' exports
+
+## 0.3.0
+
+### Minor Changes
+
+-   97dca15: Added support for service-messages!
+
+### Patch Changes
+
+-   e9e74b0: export demo mode and language related utilities
+-   Updated dependencies [97dca15]
+-   Updated dependencies [97dca15]
+    -   @equinor/mad-components@0.8.4
+    -   @equinor/mad-auth@0.1.2
+
+## 0.2.0
+
+### Minor Changes
+
+-   8e0178c: Added `AboutScreen`
+
+### Patch Changes
+
+-   bcb1f75: Now exports `authenticateSilently`
+-   Updated dependencies [59860d3]
+-   Updated dependencies [6f62312]
+    -   @equinor/mad-components@0.8.3
+    -   @equinor/mad-auth@0.1.1
+
 ## 0.1.0
 
 ### Minor Changes